requirements.txt
fastapi
uvicorn
requests
python-dotenv
langchain
langchain-google-genai   # if using Gemini
# langchain-openai        # if using OpenAI
google-generativeai 